Touch Systems Front-End React Developer Test
A project that lists Ticket Master events which are musical events, by default showing Birmingham and allowing the user to search for a different city.

It is mobile first design, controlling how it shows in a list or grid.

It uses a Modal approach to showing more information about each individual item.

An Atomic Design approach was taken for the component structure, using Atoms, Molecules and Organisms.